Markdown code block

Take a character vector of lines and return a glue vector

Usage

md_chunk(x, type = c("tick", "tilde", "indent"), ...)

Arguments

x

A character vector of lines to be wrapped concatenated into a single block, possibly created by readLines() or deparse().

type

The type of code block to be created. Either "tick", "tilde" (which call md_fence()) or "indent" (which calls md_indent()).

...

Arguments to be passed to md_fence() or md_indent().

Value

A glue object of length 1, with elements of x formatted via md_fence() or md_indent().

Details

Turn a character vector of lines into a single code block either indented or fenced in tildes or backticks. This markdown leaf block can be rendered as nested HTML <code> and <pre> tags. This function either calls md_fence() or md_indent() based on the type argument.

See also

Examples

md_chunk("$ sudo apt install r-base-dev", info = "bash")
#> ```bash
#> $ sudo apt install r-base-dev
#> ```
md_indent(
  n = c(4, 4, 6),
  x = c(
    "library(dplyr)",
    "starwars %>%",
    "filter(species == 'Droid')"
  )
)
#>     library(dplyr)
#>     starwars %>%
#>       filter(species == 'Droid')