Markdown italic emphasis

Take a character vector and wrap each element in single underscores to create a glue vector of italic emphasis text. This inline is rendered as the <em> HTML tag.

Usage

md_italic(x)

Arguments

x

The text to be emphasized in italics.

Value

A glue vector of length equal to x.

Details

A single * or _ can open or close emphasis... Emphasis begins with a delimiter that can open emphasis and ends with a delimiter that can close emphasis, and that uses the same character (_ or *) as the opening delimiter.

See also

Examples

md_italic("Example")
#> _Example_
md_italic(state.name)
#> _Alabama_
#> _Alaska_
#> _Arizona_
#> _Arkansas_
#> _California_
#> _Colorado_
#> _Connecticut_
#> _Delaware_
#> _Florida_
#> _Georgia_
#> _Hawaii_
#> _Idaho_
#> _Illinois_
#> _Indiana_
#> _Iowa_
#> _Kansas_
#> _Kentucky_
#> _Louisiana_
#> _Maine_
#> _Maryland_
#> _Massachusetts_
#> _Michigan_
#> _Minnesota_
#> _Mississippi_
#> _Missouri_
#> _Montana_
#> _Nebraska_
#> _Nevada_
#> _New Hampshire_
#> _New Jersey_
#> _New Mexico_
#> _New York_
#> _North Carolina_
#> _North Dakota_
#> _Ohio_
#> _Oklahoma_
#> _Oregon_
#> _Pennsylvania_
#> _Rhode Island_
#> _South Carolina_
#> _South Dakota_
#> _Tennessee_
#> _Texas_
#> _Utah_
#> _Vermont_
#> _Virginia_
#> _Washington_
#> _West Virginia_
#> _Wisconsin_
#> _Wyoming_