Markdown paragraphs breaks

Take a character vector and return a glue vector of paragraphs separated by double newlines. This leaf block is rendered as distinct <p> HTML tags.

Usage

md_paragraph(...)

Arguments

...

Any number of character vectors.

Value

A glue vector with elements of ... separated by two newlines.

Details

A sequence of non-blank lines that cannot be interpreted as other kinds of blocks forms a paragraph. The contents of the paragraph are the result of parsing the paragraph’s raw content as inlines. The paragraph’s raw content is formed by concatenating the lines and removing initial and final whitespace... Paragraphs can contain multiple lines, but no blank lines.

See also

Examples

md_paragraph(stringr::sentences[1:3])
#> The birch canoe slid on the smooth planks.
#> 
#> Glue the sheet to the dark blue background.
#> 
#> It's easy to tell the depth of a well.
#>