Markdown soft line breaks

Take a character vector and return a glue vector of separated by a single newline. This inline is rendered as single <p> HTML tags.

md_softline(...)

Arguments

...
Any number of character vectors.

Value

A glue vector with elements of ... separated by a single newline.

Details

A regular line break (not in a code span or HTML tag) that is not preceded by two or more spaces or a backslash is parsed as a softbreak. (A softbreak may be rendered in HTML either as a line ending or as a space. The result will be the same in browsers. In the examples here, a line ending will be used.)

See also

Examples

# compare the following
md_bold(c("One", "Two"))

#> **One**
#> **Two**

md_softline(md_bold(c("One", "Two")))

#> **One**
#> **Two**