Markdown bold emphasis

Take a character vector and wrap each element in double asterisks to create a glue vector of bold emphasis text. This inline is rendered as the <stong> HTML tag.

md_bold(x)

Arguments

x
The text to be emphasized in bold.

Value

A glue vector of length equal to x.

Details

A double ** or __ can open or close emphasis... Emphasis begins with a delimiter that can open emphasis and ends with a delimiter that can close emphasis, and that uses the same character (__ or **) as the opening delimiter.

See also

Examples

md_bold("Example")

#> **Example**

md_bold(state.name)

#> **Alabama**
#> **Alaska**
#> **Arizona**
#> **Arkansas**
#> **California**
#> **Colorado**
#> **Connecticut**
#> **Delaware**
#> **Florida**
#> **Georgia**
#> **Hawaii**
#> **Idaho**
#> **Illinois**
#> **Indiana**
#> **Iowa**
#> **Kansas**
#> **Kentucky**
#> **Louisiana**
#> **Maine**
#> **Maryland**
#> **Massachusetts**
#> **Michigan**
#> **Minnesota**
#> **Mississippi**
#> **Missouri**
#> **Montana**
#> **Nebraska**
#> **Nevada**
#> **New Hampshire**
#> **New Jersey**
#> **New Mexico**
#> **New York**
#> **North Carolina**
#> **North Dakota**
#> **Ohio**
#> **Oklahoma**
#> **Oregon**
#> **Pennsylvania**
#> **Rhode Island**
#> **South Carolina**
#> **South Dakota**
#> **Tennessee**
#> **Texas**
#> **Utah**
#> **Vermont**
#> **Virginia**
#> **Washington**
#> **West Virginia**
#> **Wisconsin**
#> **Wyoming**