Markdown Code Block

Wrap a character vector (possibly lines of a file) in code fences to be rendered as a code block.

md_chunk(x, lang = "r", type = c("ticks", "tilde", "indent"),
  cat = TRUE)

Arguments

x
A character vector of lines (with leading spaces for indentation) to be wrapped in fences.

lang
The language of the code within the fence. Aside from clarity, this allows many markdown engines to adjust the syntax highlighting. Defaults to "r". This option is not used with indented blocks.

type
The type of code block to be created. Defaults to tipple backticks code fences, but you can also use 4-spaces (useful for legacy support on Reddit).

cat
logical; Should the list be concatenated and printed, with each bullet element separated by a new line? Defaults to TRUE.

Value

A character vector wrapped on either side by code fences.

Details

From the GFM spec: A code fence is a sequence of at least three consecutive backtick characters… A fenced code block begins with a code fence, indented no more than three spaces… The line with the opening code fence may optionally contain some text following the code fence… The content of the code block consists of all subsequent lines, until a closing code fence… The content of a code fence is treated as literal text, not parsed as inlines. The first word of the info string is typically used to specify the language of the code sample

Examples

md_chunk(c("library(ggplot2)", "ggplot(mpg) +", "  geom_point((displ, hwy))"))

#> ```r
#> library(ggplot2)
#> ggplot(mpg) +
#>   geom_point((displ, hwy))
#> ```