Convert markdown to HTML

Take a character vector of valid markdown text and pass it to markdown::markdownToHTML() to create a glue vector of HTML fragments. Primarily used to test that md_*() functions create vectors that meet the GFM spec and can be rendered as HTML.

md_convert(x, frag = TRUE, disallow = TRUE)

Arguments

x

A character vector of markdown text to be converted.

frag

logical; Whether only a single HTML fragment should be returned. TRUE by default.

disallow

logical; Should md_disallow() be called on the converted output?

Value

A glue vector of length 1 containing HTML tags.

Details

GFM enables the tagfilter extension, where the following HTML tags will be filtered when rendering HTML output...

See also

Examples

md_convert(x = md_bold("test"))
#> <p><strong>test</strong></p>
#>