Markdown hard line breaks

Take a character vector and return a collapsed glue vector with each original element separated by two spaces and a newline. This inline is rendered with a <br /> HTML tag.

md_hardline(...)

Arguments

...

Any number of character vectors.

Value

A glue vector with elements of ... separated by two trailing spaces and a single newline.

Details

A line break (not in a code span or HTML tag) that is preceded by two or more spaces and does not occur at the end of a block is parsed as a hard line break (rendered in HTML as a <br /> tag)

See also

Examples

# compare the following
md_bold(c("One", "Two"))
#> **One**
#> **Two**
md_hardline(md_bold(c("One", "Two")), md_italic("Three"))
#> **One**  
#> **Two**  
#> _Three_